+++ packages/wrapper/src/editorAppClassic.ts
@@ -4,9 +4,9 @@
  * ------------------------------------------------------------------------------------------ */
 
 import * as monaco from 'monaco-editor';
+import { Logger } from 'monaco-languageclient/tools';
 import { EditorAppBase, EditorAppConfigBase, ModelUpdateType, isEqual, isModelUpdateRequired } from './editorAppBase.js';
 import { UserConfig } from './wrapper.js';
-import { Logger } from './logger.js';
 
 export type EditorAppConfigClassic = EditorAppConfigBase & {
     $type: 'classic';